updated express api
setup testing with jest and supertest
