+++ package.json
@@ -9,7 +9,8 @@
     "start": "node dist/index.js",
     "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/index.js\"",
     "db:migrate": "cd dist && knex migrate:latest",
-    "db:seed": "cd dist && knex seed:run"
+    "db:seed": "cd dist && knex seed:run",
+    "test": "cross-env NODE_ENV=development jest --testTimeout=10000"
   },
   "repository": {
     "type": "git",
@@ -22,6 +23,7 @@
   },
   "homepage": "https://github.com/JohnRamberger/dsgt-member-portal#readme",
   "dependencies": {
+    "@types/supertest": "^2.0.12",
     "concurrently": "^7.2.2",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
@@ -34,11 +36,16 @@
     "pg": "^8.7.3",
     "qrcode": "^1.5.1",
     "request-ip": "^3.3.0",
-    "stripe": "^9.16.0"
+    "stripe": "^9.16.0",
+    "ts-jest": "^29.0.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
+    "@types/jest": "^29.0.3",
     "@types/node": "^18.0.3",
+    "cross-env": "^7.0.3",
+    "jest": "^29.0.3",
+    "supertest": "^6.2.4",
     "ts-node": "^8.10.2",
     "typescript": "^4.7.4"
   },
